net-snmp: Fix for Perl 5.26
This adds 2 patches pulled from upstream, and then autospec is rerun.

The only failures in the internal test suite are caused by the lack of
DES. This is a bug in the build process, it should detect that DES is
not configured and update its help messages and the set of tests to
run to reflect that DES is not configred.

The first changes "U64" to "struct counter64", as the new version of
perl defines its own U64 type.

The second works around an internal perl bug. Essentially the "@_"
array which is used to pass parameters to subroutines is not correctly
protected against garbage collection. This leads to values being freed
prematurly, and then perl panics when it tries to use them. Some of
the gruesome details are here.
https://rt.perl.org/Public/Bug/Display.html?id=125907#txn-1363270
where it shows values on the stack being shown incorrectly.

Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>